chore: Fix typos and clean-up unused test inputs

+++ test/test_ssh_tunnelling_with_docker.sh
@@ -5,21 +5,25 @@
 source "../distcc.sh"
 
 
-_no_command() {
+_command() {
   if ! command -v "$1" &>/dev/null; then
     echo "SKIPPING: '""$1""' is not available!" >&2
-    return 0
+    return 1
   fi
-  return 1
+  return 0
 }
 
-skip_if "_no_command docker || _no_command netcat || _no_command ssh" "test"
+skip_if "! _command docker || ! _command netcat || ! _command ssh" "test"
 
 
 IMAGE="distcc-driver-ssh-test"
 CONTAINER="distcc-driver-ssh-test-test-1"
 
 setup_suite() {
+  if ! _command "docker"; then
+    return
+  fi
+
   # Build and start a Docker container.
   docker build \
     -t "$IMAGE" \
@@ -37,6 +41,10 @@ setup_suite() {
 }
 
 teardown_suite() {
+  if ! _command "docker"; then
+    return
+  fi
+
   # Clean up the Docker artefacts.
   docker kill "$CONTAINER"
   docker rm "$CONTAINER"
